Don't mange section layout in event module widgets
